backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pillow
requests
pydantic
python-dotenv